Added NurbsBase abstract class. NurbsCurve inheritance the NurbsBase. Unification of the abstract class across the project.
