+++ src/GShark/Intersection/BoundingBoxTree/LazyCurveBBT.cs
@@ -6,13 +6,13 @@ using GShark.Interfaces;
 
 namespace GShark.Intersection.BoundingBoxTree
 {
-    internal class LazyCurveBBT : IBoundingBoxTree<NurbsCurve>
+    internal class LazyCurveBBT : IBoundingBoxTree<NurbsBase>
     {
-        private readonly NurbsCurve _curve;
+        private readonly NurbsBase _curve;
         private readonly BoundingBox _boundingBox;
         private readonly double _knotTolerance;
 
-        internal LazyCurveBBT(NurbsCurve curve, double knotTolerance = double.NaN)
+        internal LazyCurveBBT(NurbsBase curve, double knotTolerance = double.NaN)
         {
             _curve = curve;
             _boundingBox = new BoundingBox(curve.ControlPointLocations);
@@ -30,18 +30,18 @@ namespace GShark.Intersection.BoundingBoxTree
             return _boundingBox;
         }
 
-        public Tuple<IBoundingBoxTree<NurbsCurve>, IBoundingBoxTree<NurbsCurve>> Split()
+        public Tuple<IBoundingBoxTree<NurbsBase>, IBoundingBoxTree<NurbsBase>> Split()
         {
             Random r = new Random();
             double t = (_curve.Knots[_curve.Knots.Count - 1] + _curve.Knots[0]) / 2.0 +
                        (_curve.Knots[_curve.Knots.Count - 1] - _curve.Knots[0]) * 0.1 * r.NextDouble();
-            List<NurbsCurve> curves = _curve.SplitAt(t);
+            List<NurbsBase> curves = _curve.SplitAt(t);
 
-            return new Tuple<IBoundingBoxTree<NurbsCurve>, IBoundingBoxTree<NurbsCurve>>
+            return new Tuple<IBoundingBoxTree<NurbsBase>, IBoundingBoxTree<NurbsBase>>
                 (new LazyCurveBBT(curves[0], _knotTolerance), new LazyCurveBBT(curves[1], _knotTolerance));
         }
 
-        public NurbsCurve Yield()
+        public NurbsBase Yield()
         {
             return _curve;
         }
