+++ src/GShark/Optimization/CurvePlaneIntersectionObjectives.cs
@@ -8,7 +8,7 @@ namespace GShark.Optimization
     /// </summary>
     public class CurvePlaneIntersectionObjectives : IObjectiveFunction
     {
-        private readonly NurbsCurve _curve;
+        private readonly NurbsBase _curve;
         private readonly Plane _plane;
 
         /// <summary>
@@ -16,7 +16,7 @@ namespace GShark.Optimization
         /// </summary>
         /// <param name="crv">ICurve used in the intersection process.</param>
         /// <param name="plane">Plane used in the intersection process.</param>
-        public CurvePlaneIntersectionObjectives(NurbsCurve crv, Plane plane)
+        public CurvePlaneIntersectionObjectives(NurbsBase crv, Plane plane)
         {
             _curve = crv;
             _plane = plane;
